qw/requirements.txt
altgraph==0.17.2
async-generator==1.10
attrs==22.1.0
beautifulsoup4==4.11.1
bottle==0.12.23
bottle-websocket==0.2.9
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
Eel==0.14.0
emoji==2.0.0
future==0.18.2
gevent==21.12.0
gevent-websocket==0.10.1
greenlet==1.1.2
h11==0.13.0
lxml==4.9.1
outcome==1.2.0
pefile==2022.5.30
pycparser==2.21
pyinstaller==5.3
pyinstaller-hooks-contrib==2022.8
pyparsing==3.0.9
PySocks==1.7.1
python-dotenv==0.20.0
pywin32-ctypes==0.2.0
requests==2.28.1
selenium==4.4.3
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
trio==0.21.0
trio-websocket==0.9.2
urllib3==1.26.11
whichcraft==0.6.1
wsproto==1.1.0
zope.event==4.5.0
zope.interface==5.4.0